LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-base-element/base-data-expected.txt

CONSOLE MESSAGE: Blocked setting data:/,test as the base URL because it does not have an allowed scheme.
CONSOLE MESSAGE: Blocked setting data:/,test as the base URL because it does not have an allowed scheme.
CONSOLE MESSAGE: Blocked setting data:/,more-test as the base URL because it does not have an allowed scheme.

PASS First <base> has a data: URL so fallback is used
PASS First <base> is removed so second is used
PASS Dynamically inserted first <base> has a data: URL so fallback is used
